Updated react app to look and act like angular and iframe sample apps Added EPL2 header for files that were missing them.
