+++ .gitignore
@@ -8,7 +8,7 @@
 # See https://help.github.com/ignore-files/ for more about ignoring files.
 
 # dependencies
-/node_modules
+/webClient/node_modules
 
 # testing
 /coverage
