--- package.json
@@ -1,32 +1,0 @@
-{
-  "name": "react-sample-app",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "react": "~16.4.0",
-    "react-dom": "~16.4.0",
-    "react-router": "~4.3.1",
-    "react-scripts-ts": "2.16.0",
-    "webpack-cli": "~3.0.0"
-  },
-  "scripts": {
-    "start": "react-scripts-ts start",
-    "build": "webpack --progress --colors",
-    "test": "react-scripts-ts test --env=jsdom",
-    "eject": "react-scripts-ts eject"
-  },
-  "devDependencies": {
-    "@types/jest": "~23.0.2",
-    "@types/node": "~10.3.2",
-    "@types/react": "~16.3.17",
-    "@types/react-dom": "~16.0.6",
-    "css-loader": "~1.0.0",
-    "html-loader": "~0.5.5",
-    "html-webpack-plugin": "~3.2.0",
-    "exports-loader": "~0.6.3",
-    "typescript": "~2.9.1",
-    "ts-loader": "~4.4.2",
-    "webpack": "~4.0.0",
-    "webpack-config": "~7.2.1"
-  }
-}
