--- tsconfig.json
@@ -1,6 +1,0 @@
-{
-  "extends": "../zlux-app-manager/virtual-desktop/plugin-config/tsconfig.react.base.json",
-  "compilerOptions": {
-    "skipLibCheck": true
-  }
-}
