--- tslint.json
@@ -1,5 +1,0 @@
-{
-  "extends": [
-    "../zlux-app-manager/virtual-desktop/plugin-config/tslint.base.json"
-  ]
-}
