+++ webClient/package.json
@@ -1,0 +1,37 @@
+{
+  "name": "org.zowe.zlux.sample.react",
+  "version": "1.0.0",
+  "description": "A sample react showcasing react framework in Zowe App Framework",
+  "license": "EPL-2.0",
+  "dependencies": {
+    "react": "~16.4.0",
+    "react-dom": "~16.4.0",
+    "react-router": "~4.3.1",
+    "react-scripts-ts": "2.16.0"
+  },
+  "scripts": {
+    "start": "webpack --progress --colors --watch",
+    "build": "webpack --progress --colors",
+    "test": "react-scripts-ts test --env=jsdom",
+    "eject": "react-scripts-ts eject"
+  },
+  "devDependencies": {
+    "@types/jest": "~23.0.2",
+    "@types/react": "~16.3.17",
+    "@types/react-dom": "~16.0.6",
+    "copy-webpack-plugin": "~4.5.2",
+    "css-loader": "~1.0.0",
+    "exports-loader": "~0.7.0",
+    "file-loader": "~1.1.11",
+    "html-loader": "~0.5.5",
+    "script-loader": "~0.7.2",
+    "source-map-loader": "~0.2.3",
+    "ts-loader": "~4.4.2",
+    "typescript": "~2.9.0",
+    "tslint": "~5.10.0",
+    "url-loader": "~1.0.1",
+    "webpack": "~4.0.0",
+    "webpack-config": "~7.5.0",
+    "webpack-cli": "~3.0.0"
+  }
+}
