+++ webClient/src/App.css
@@ -1,4 +1,5 @@
 
+
 /*
   This program and the accompanying materials are
   made available under the terms of the Eclipse Public License v2.0 which accompanies
@@ -12,93 +13,109 @@
   font-family: sans-serif;
   font-size: 0.8em;
 }
- .bold-text {
+
+.bold-text {
   font-weight: 600;
 }
- .bigger-bold-text {
+
+.bigger-bold-text {
   font-size: 1.1em;
   font-weight: 600;
 }
- .div-input {
+
+.div-input {
   margin-top: 5px;
   margin-bottom: 10px;
   padding-top: 5px;
   padding-bottom: 5px;
   width: 300px;
 }
- .div-textarea-input {
+
+.div-textarea-input {
   margin-top: 5px;
   margin-bottom: 5px;
   padding-top: 5px;
   padding-bottom: 5px;
   width: 95%;
 }
- .iframe-input {
+
+.iframe-input {
   width: 100%;
   padding-left: 5px;
 }
- .input-corner {
+
+.input-corner {
   border-radius: 4px;
 }
- .input-height {
+
+.input-height {
   height: 20px;
 }
- .input-text {
+
+.input-text {
 }
- .iframe-button {
+
+.iframe-button {
   border-radius: 4px;
   float: right;
   margin-right: 30%;
   font-weight: 700;
 }
- .right-align {
+
+.right-align {
   text-align: right;
   padding-right: 50px;
 }
- .shadowed {
+
+.shadowed {
   box-shadow: 3px 3px 10px;
 }
- .hide-it {
+
+.hide-it {
   display: none;
 }
- .bottom-10 {
+
+.bottom-10 {
   margin-bottom: 10px;
 }
- .display-text {
+
+.display-text {
   border-radius: 4px;
   border: 1px solid grey;
   background-color: white;
   height: 18px;
   width: 96%;
 }
- .disable-effect {
+
+.disable-effect {
   color: grey;
 }
- h1 {
+
+h1 {
   text-align: center;
- }
- 
- .server-response {
-   width: 100%;
+}
+
+.server-response {
+  width: 100%;
   height: 180px;
 }
- .test-panel {
+
+.test-panel {
   padding: 5px 5px 5px 5px;
 }
- .test-panel-container {
+
+.test-panel-container {
   flex-direction: row;
   display: flex;
 }
- .plugin-test-panel {
+
+.plugin-test-panel {
   border-right: outset;
   flex-grow: 1;
 }
- .dataservice-test-panel {
-  border-left: outset;
-}
 
-.hundred-width {
-  width: 100%
+.dataservice-test-panel {
+  border-left: outset;
 }
 /*
   This program and the accompanying materials are
@@ -109,3 +126,4 @@
   
   Copyright Contributors to the Zowe Project.
 */
+
