+++ webClient/src/mvd-resources.ts
