+++ webClient/tsconfig.json
@@ -1,0 +1,6 @@
+{
+  "extends": "../../zlux-app-manager/virtual-desktop/plugin-config/tsconfig.react.base.json",
+  "compilerOptions": {
+    "skipLibCheck": true
+  }
+}
